requirements.txt
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Login==0.5.0
rake-nltk==1.0.6
Flask-Cors==3.0.10
google-cloud-storage==2.0.0
pandas==1.1.5
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-cloud-core==2.3.1
json-server==0.1.3
gcsfs==2022.1.0
gunicorn==19.9.0
google-cloud-vision==2.7.3
Werkzeug==2.0.3